# perf: remove assoc