@arnaudgeiser 2023-02-28
> I don't mind merging this, since that assoc seems a bit pointless, but I don't see any evidence it's actually faster. The variance is way too large to draw any conclusions.

I don't think it's making a huge difference.
But I can definitely see the time spent on the `assoc` on flame graphs I've generated.